[object_test] fix race in object tests
Several tests were working like:

  create object
  tick1 = ...
  wait event
  tick2 = ...

The problem is the event can happen between the creation and the wait
call, meaning tick2 - tick1 will be a lower bound of the actual
duration. We fix this by registering a handler detecting and tracking
the next object added in the model. Creation and updates can be recorded
accurately.