[setup_test] introduce ClientOpts to run client/sim/replayer
There are a couple of connect*() functions already, but the problem is
we cannot easily add options, like logging configuration. The
connectAllUserAndWait() function now takes a ClientOpts instance and a
NewAllUserOpts() constructor is added to mimic the default values
behaviour. Other connect*() functions will be refactored to use it, then
new options will be added.