[model] fix asynchronous termination upon disconnection
The input handler was closing the request channel upon termination
(usually remote link disconnection). This is wrong because the requests
channel does not belong to the handler, but is shared with main
goroutine clients. What we want is to close the channel once we know:
- the update handler will not be called anymore
- the main caller is done with the model

The simple solution is to have a WaitGroup for the update handler, wait
for it in Close() then close the channel.