[logistic_plugin] move request history index into ConsignIndex
It duplicates the requests index of LogisticPlugin but this is fine, it
contains only live requests. The good part is it regroups all the
request history updating code in a single place. All *Resolver classes
now handles only the legacy logging.

This will also help reinitializing the index in replay mode.