[logistic_plugin] add a reloading read-only mode, activate in replay
The logistic plugin is now activated in replay and reinitialized from
the logistic archive during the plugin initialization. Hopefully, this
will not be too slow and can be kept synchronous, I still need to do
some benchmarking.

At this point, I have only tested the API with gosword, gaming does not
seem to send logistic queries in replay.