[replayer][dispatcher] Fix crash replayer when a radio silence is lifted
Crash happened when an agent knowledge was destroyed inside a keyframe and the corresponding knowledge group was destroyed in one of the following ticks but before the end of the syncrhonisation process. The fix is to delay all object destructions during ticks which follow the key frame to the end synchronisation step.