[logistic_plugin] writes a max valid offset file in ConsignArchive::Flush()
The replayer can be run on a dataset which is being written by a running
simulation. Since concurrent read/write have no atomicity guarantees, we
need a way to signal which part of the requests archive file is valid. A
secondary "available" file containing the index of the last file and the
offset of the last valid byte within this file is written in Flush() and
when files are being rotated.

To ensure this file can be read safely, the information is first written
in a temporary file, then target file is removed and replaced by the
temporary file. This leaves a window where the offset file does not
exist but this is better than reading invalid data.