[logistic_plugin] periodically refresh the data store in read-only mode
In read-only mode, the ConsignRecorder checks the archive "offset" file
and possibly load new entries. The store is refreshed at most every 5
seconds (this value is currently hard coded but might become
configurable for tests purpose).

This feature is required to handle the case of a replayer reading data
generated by a live simulation.