[simu] move ListDmpFiles in errors.go, add CheckSessionErrors
We want to check for errors after successful test runs. For instance,
some component may crash and the crash may be unnoticed, or a failure
might not clearly be a crash without investigating.

The CheckSessionErrors() function will regroup all kind of checks from
dump file, log file and so forth.