[setup_test] add AddGaming() helper to start gaming in test
AddGaming() is used on *ClientOpts like AddLogger, calls may be chained.
It stops the simulation, starts a gaming and wait for one tick to pass
before proceeding with the test.

It is not perfect, if you pause and close gaming, you have to wait for
the test to timeout and/or cleanup some stray process, but is makes
visualizing tests much easier.