[crowd_test] add report logger, fix crowd checkpoint test
The previous version was passing but the crowd was never stopped because
there were not enough units to stop it. The new version fixes it and
captures the reports emitted by the crowd being stopped at the
checkpoint.

To do this, a Reporter struct is added, recording all reports maching a
supplied filter between Start() and Stop() calls. Helpers are provided
to generate filters for report types or matching reports.xml messages.