[meteo] MeteoLocal was sending invalid cloud density (SWBUG-11627)
The protobuf code is duplicated between the dispatcher and the
simulation (see PHY_LocalMeteo) which explains why local weather worked
in gaming but were not restored after a checkpoint.

The duplication is not easily fixed as PHY_LocalMeteo inherits from
Meteo and not MeteoLocal. Unifying PHY_LocalMeteo and MeteoLocal is
probably the way to go.