Revert "[client_gui,clients_kernel,gaming_app] removed tools::findTranslation which was a duplicate of tools::translate"
This reverts commit e7b0d38eb3cf06126e8485390ac093e46e2470f6.

The motivation is findTranslation is different from translate as it is
not picked by Qt tools, and relies on the existence of a translation
with the same context in another file. This behaviour is brittle but
removing it means translations have to be duplicated in other ts files
which should be at least partially automated.