[unitmagicaction_test] pass knowledge instead of carrier id to Helitransport
The physical db declares the parameter as LocationComposite restricted
to AgentKnowledge. But the simulation does not care and happily returns
a successful ack and a mission. Unfortunately, recent changes to
decisional make the mission to fail.

Instead we pass an AgentKnowledge parameter but the trick is the
knowledge is filled with the agent identifier.