[simulation_kernel] SWBUG-11688 rewrote change equipment state magic action
The idea is instead to blindly overwrite equipment states in arbitrary order to try and leave the existing states as untouched as possible.
So we start by skipping the states which match exactly between the current and the new state list, then we update only the breakdown of matching states and only finally we overwrite the remaining states.