[simulation] Prevent flows from flowing into their source concentration, from being blocked by non-blocking objects and from leaving their concentration to continue by themselves when no longer blocked.
Linked to JIRA # SWBUG-11649 Pas de creation de connaissance quand une foule est bloquee dans un objet
SWBUG-10493 : When a crowd runs into a barricade, it must not only output a report but also acquire the knowledge of the object.
TestCrowdInCheckpoint
(The latter must be enriched with an additional barricade before the crowd reaches the checkpoint).