[adaptation] SWBUG-11533: ALGERIE - plantage adaptation à la sauvegarde
To reproduce the crash : open a database, choose mission tab, select a mission, clic in the drawing combo box, close the database.

It's a problem within the Qt deletion process, dues to active focus and related to our Drawings since it does not happen on other ComboBoxes.

This is fixed with a call to clear() in the MainTabWidget destructor, before letting QTabWidget's destructor handle it's children deletion.
The clear() method removes the children widgets visually, but it lets the ownership to the QTabWidget so they will still be deleted by him, but with no active focus on any of them.