[simulation] Fix JIRA # SWBUG-11856 Simulation crash during random tests (02/20/2014)
Crash happened because interaction_ of MIL_Object was not cleaned soon enough:
magic action: object was marked for destruction.
First Tick: Object was still in interaction_, so NotifyExternalPerception was called.
interaction_ was cleaned. Unit was deleted.
Second Tick: The NotifyExternalPerception pointer kept from previous tick was used : Crash.

I leave the existing call to Clean because it's needed to ensure we don't keep pointers to freed memory when saving an automatic checkpoint right afterwards. It may be overkill because the decisional keeps a pointer to the object, but I think it's a bit brittle to rely on. Another fix would be to defer the destruction of the unit in case of magic action into the Unit's UpdateState method, but we would have to check all the side effects...