[unitmagicaction_test] fix TestUnitChangeHumanState after sim change
Partial updates performed by change_human_state are ambiguous and should
probably be forbidden by the simulation. The problem is if a unit has 2
troopers and you send a message referencing only one, there is no way to
know which trooper it applies to. The previous simulation code was
mixing states more or less randomly while the new one:
1. Match unit existing states by rank and remove them from the request
2. Applies the remaining states

Assuming a unit with 2 healthy troopers, running the following magic
actions in sequence:
- Make 1 trooper injured
- Make 1 trooper healthy
Now always result in a unit with 1 healthy trooper and 1 injured
trooper, the second request being turned in a no-operation.

This patch resets the state of all troopers in the test.