[crowd_test] remove Pause/Resume from TestCrowdTeleportation, fix test
The test was silently timing out in the step() command but succeeded
anyway by construction. I suspect some mismatch in the Resume and
expected elapsed ticks calculation.

The test was rewritten without pause/resume, with the assumption the
crowd will not end its movement before we detect it starts moving and we
teleport it. If it happens, we could just give it a very long path.

This reduces test run time from 60s to 4/5s.