[handlers_sim] extract tick from ControlInformation only in initial state
And applies a -1 offset on the value.

The startup situation is still confused. What happened in
TestReplayerModels was the client seeing a ControlInformation with
tick=2, setting the Model.Tick to 2 then resuming(1) expecting to reach
tick 3. Unfortunately, the replayer thought otherwise and gave it tick 2
again.

This reduces TestReplayerModels run time from 70s to 8s.