[crowd_test] move barricade position a bit to fix TestCrowdInCheckpoint
I am not happy with the fix but Laurent already fought several days with
crowds and object tolerances and I have no intention to spend more time
on this now. The problem is the crowd now "detects" the object and
avoids it, though it should not be possible at all since the object
belongs to another party.