[actions_gui] SWBUG-11970: prevent QTextEdit to emit textEdited signal when we force focus on another QTextEdit.
The problem was the following:
Calling QLineEdit::setFocus emits textEdited signals "sometimes", and doesn't "the other times".
According to qt's documentation, this signal is supposted to be emitted whenever the text is edited,
and not emitted when the text is changed programmatically.

After a quick investigation of qt's sources, it behave like that:
- If the QLineEdit is part of the active windows, QApplication remove the focus of the previously focused widget.
Doing so on a QLineEdit calls deselect() (which emits the textEdited signal) if the reason of the new focus is
anything else than Qt::ActiveWindowFocusReason or Qt::PopupFocusReason.
- If it's not part of the active windows, QApplication does nothing.

That's why the reproductibility wasn't easy, it depended on the docked status of the MissionPanel,
cause the QLineEdit was part of the active windows if the panel was docked.

So, forcing the new focus with Qt::ActiveWindowFocusReason as reason ensure that the signal won't be emitted,
whenever the panel is docked or not.