[gaming] SWBUG-11868 all requests in a 'finished' state are now displayed only in the "Show completed requests" mode
Until now a request had to be destroyed to be moved to the completed request list.
This new requirement is mainly for medical requests as some of them are never destroyed but should still be considered completed.
It however probably doesn't hurt to do this for all requests as the finished state update message is very often sent together with the destruction message anyway.