[simulation_kernel] fixed crash when an exception happens during the un-serialization of a logistic role
Boost.Serialization destroys any partially loaded object when an exception happens during un-serialization.
Therefore we need to wait for a role to be fully un-serialized before registering it as this gives ownership to the role container which will also try and delete the role when destroyed.
And double destruction is bad.