[simulation] Fix JIRA # SWBUG-12208 [NZ] Checkpoints corrompus
Checkpoints were corrupted because a convoy unit was destroyed (through a breakdown or otehrwise).
The SupplyConsign was destroyed but remained associated to the destroyed unit, and it pointed to invalid memory.
The change in MIL_AutomateLOG::SupplyDestroyConvoyPion may be enough to fix the issue but SupplyConvoyReal::Finish almost certainly didn't do what was expected when receiving false.