[simulation_kernel] fixed crash when an exception is raised during the de-serialization of a unit
The problem was that the units were loaded before the armies so when an exception occurred armies were deleted first, but ~Formation and ~MIL_KnowledgeGroup call methods on their Army which resulted in a crash.
The issue is a bit difficult to test because it means either tamper with the serialized checkpoint or change the code to throw an exception (for instance from PHY_RolePionLOG_Maintenance::load).