[build] pull sword current protobuf and compare them here
It inverses the sword / sword-proto dependency which is nice because it
means we no longer have to version sword-proto. The drawback is we may
take more time to detect protobuf compatibility issues. But this is not
a big problem in practice as we only need to fix them before releases.