[simulation_kernel] SWBUG-12286 fixed (some) pathfind threading issues
The problem was that a DEC_PathFindRequest was the last owner of a shared path which was being destroyed in a worker thread.
To prevent this we use a weak_ptr instead along with the feature that acquiring a shared_ptr with lock() is thread-safe.
A nice side effect is that we then don't need manual reference counting and associated job canceling anymore.