[clients_gui] fix WGS84 formatting for small values (SWBUG-12361)
The previous code was checking degree sign to decide N/S or E/W suffix.
This did not work when negative coordinates were rounded up to zero.
Instead we track existing prefixes or a negative sign and reuse it
later.

Note the test is a bit clunky because the DMS formatting string is not
valid UTF-8 string which causes a lot of issues when converting to and
from QString. But I prefer not to touch this now because the degree
symbol may be embedded in the code in a lot of places.