[gaming_app] fixed a bug where a waypoint could not be dragged or removed when part of an incomplete path
This tends to happen a lot when adding a waypoint to an already existing path and dragging it to the border of the map (for instance in logistic_example).
I believe the fact that the path cannot be computed is a bug as adding more waypoints and/or removing some of them usually works around it, but this is an issue for another day.