[simulation_kernel] fixed serialization warning
BOOST_CLASS_EXPORT is supposed to force template instantiation when (un)serializing through a base class pointer. Therefore Boost.Serialization checks that the class of the serialized object is "polymorphic".
My guess is it only triggers in vc100 because mpl::print might be broken in x64 MSVC (couldn't find a ticket though).
The quickest fix is to slap a virtual on the class destructor. A more proper fix would be use concrete archive types in save/load.