[fire_test] fix TestActiveProtection (SWBUG-12467)
I am not completely sure why the fix works, I suspect there is a race
between the unit attribute update and the resource consumption caused by
the active protection. But I ran it at least 100 times in a noisy
environment and it did not fail while it previously did after 15
attempts.