[tools] emit ConnectionSuccess event before reading data
What motivates this change is TestModelInitialization was apparently
failing because sometimes a client connection authentication key request
message was refused because the connection was not registered in
ClientsNetworker. One way to reach this state was simply to process the
authentication key request before handling the connection success event.
Looking at SocketManager code, the connection sequence was like:

  s = CreateSocket()
  s.StartReading()
  callback.ConnectionSuccess()

But StartReading is a boost::asio call and this might result in the
following events:

  s = CreateSocket()
  s.StartReading()
 [in other thread] Read() => MessageCallback_ABC.OnMessage()
  callback.ConnectionSuccess()

This fix just swaps the connection success notification and
StartReading().