[replayer] handle ControlStop message, implement in gosword
For some reason the previous dispatcher/simulation changes caused
replayer tests to fail because the listening address/port was still
used. I am certain this is restricted to replayer tests as only those
fails if the listening port is singled out. Supporting ControlStop
message and using it in gosword when terminating the replayer seem to
solve the problem.