[schemas] fix 5.3.3 -> 5.4.0 breakdowns XSL (SWBUG-12569)
The error is to create the child element before evaluating parent
element attributes. The XSL processor emits:

  XSLT warning: An attribute cannot be added to the result tree when
  there is no open element. The attribute will be ignored. (Occurred in
  entity 'schemas/5.4.0/xsl/breakdowns_5.3.3-5.4.0.xsl', at line 16, column 16.)

which actually means that *parent* attributes are dropped, while the
child element is created. breakdowns.xsd declares no attribute for its
root element, this error is probably harmless but for warnings when
migrating physical databases, as long as we used this specific XSL
processor.