[timeline_server] unify global and dedicated logger
For some reason lost in merged git repositories history, the log module
was being shadowed by another logger, the latter being passed to the
controller and web server. Unify both instead.

I created a single logger instance to avoid refactoring the controller
and web server loggers. It is probably done that way to easily mock the
logger in tests.