[adn] do not record max-repair-duration=0s (SWBUG-12389)
The simulation refuses to load max-repair-duration=0s values. I do not
want to allow these values now because we have no reason to. And I do
not want to interpret 0s as +inf because 0s is a meaningful input, only
useless *for now*.

Instead, make ADN not write the value if it is set to 0s. Note this is
not completely new since ADN would *read* it as unset if set to 0s.