[simtests] avoid race in TestLogisticDeployment
The problem with the WaitTicks(10) following the deployment mission is
it gives no guarantee the unit will have deployed when done. A more
reliable approach is to count received reports before the mission,
assume none will be received between the Count() and the mission (no
starting deploying report should be emitted after the TC2 declares
itself deployed). Then order the TC2 to deploy and wait until a new
"deployment done" report is received.

This version was successfully run in loop while compiling a simulation
with -j8, while the previous version would fail after a dozen of loops.