[gotester] run one process per test case, dispatch dynamically
Instead of splitting the test set into N groups run in N child process,
we extract the test package along with its name, compile the package
test executables and use them to run one test per process. The advantage
is it improves the dispatching of tests over process. The disadvantage
is it spawns a lot of process which Windows is not really good at.
Still, it is faster than the previous mode:

  Before (best of 3): 5mn19s
  After  (best of 3): 4mn37s

Width -Dninja.jobs=10 instead of 8, run time reduces to 4mn24 and
process monitoring shows we are much closer to saturating the CPUs.