[gaming] work around knowledge update freeze (SWBUG-12850)
The Brasilian exercise generates a log of agent knowledge updates which
trigger costly updates in gaming. This fixes reduces the freeze to
almost nothing instead of a long one of one minute or more.
The culprit is at least StandardModel::FindDataItem() and possibly the
following invalidate(). Note the FindDataItem() performs full recursive
traversal of tactical hierarchies or trees of similar size, for each
agent knowledge update, which are often unneeded.

There is a lot of room for improvements here:
- Avoid unnecessary notifications, like NotifyUpdate( Entity_ABC )
- Fix StandardModel::FindItemData, or specialize it for trees
- Remove invalidate() calls