[simulation] Fix JIRA # SWBUG-12774 [Angers] Ambiance pédagogique
Partial revert of a refactor which broke speed modulations.
rSpeedModificator_ must be taken into action when moving.
Theoric and isMax are totally separate concepts. One is maximum speed modified by nothing. The other is modified by rMaxSpeedModificator_. rSpeedModificator_ must be considered when asking for the current speed.
The modes are: theoric max speed, current max speed and current speed (decisional dependent). The 4th case is theoric current speed which is used in pathfind - the only effect of using that instead of current speed is likely the effect of traffic.