[simulation] Fix JIRA # SWBUG-12895 Perte de connaissance d'une unité capturée que l'on vient de déposer
Transported units must be warned when they enter or leave objects (in particular urban blocks).
However, we must not cause an explosion if a transported unit moves on a landmine as the transporter already handles attrition.
Avoid reentrance in case a unit is transporting another unit that reinforces it or vice versa.