[sim] rewrite MIL_UrbanObject scaled cache with map+list
And fix the bug about the stale entries not being evicted from the map.

I benchmarked the GetScaledLocation() in best case and worse case (100%
cache hit and 100% cache miss) and it consistently beats the previous
version. Interestingly, it is also faster than:
- single vector with erase + insert
- single deque with erase + insert
- unordered_map + list

The only sad part is I have no idea how to write it in such a way the
list holds a map iterator instead of a key which would save one lookup.
The typedef eludes me.