[simulation] Removal if duplicate DEC functions, made all functions which could be accessible to automata and agents and crowds
There is one bug fix inside : DEC_KnowledgeAgent_IsPerceptionLevelMax was registered at archetype level with a GetAutomate() reference. This ment when called by another automat, it was always called in the context of the first created automat.
There is one change in behavior : Crowds used to consider an object knowledge invalid if the underlying object was destroyed - even though they didn't know it. Now crowds, agents, automata all behave the same: a knowledge is valid if the simulation says so. It will only be considered invalid by the decisional if the object is destroyed and the entity knows it (i.e. if the simulation considers it invalid).