[simulation] Make it easier to debug serialization.
There is a BOOST_CLASS_EXPORT_KEY but not implement for ( sword::Sink_ABC ). The class being abstract, both are useless.
Chained & and >>/<< make it hard to understand the stack when the serialiation crashes.
Serializing id before pointer makes it more likely to be able to find out which agent knowledge was causing an issue when deserializing after an access violation happened on save during automatic testing.
Avoid a crash in DEC_KS_ObjectInteraction::Talk : if the interaction is not active the tick when an object is destroyed, then we will crash later.