[gaming] rewrite WeaponRangeLayer without shaders (SWBUG-13128)
The problem was the shader uniforms size was depending on the input data
size and was quickly hitting video card limits. Instead, the layer is
rewritten to merge shapes belonging to the same color group together to
avoid blending effects, then blend the different color groups together.
Geos is used for the geometry union computations and GLU tesselator to
generate OpenGL compatible features.