[simulation_kernel] restore check to GetObjectKnown()
This check was lost in a refactoring. It's actually stronger than IsValid because whenever an object is invalidated, the known associated object is also reset.
It's a little odd as this means a unit will not compute a new path around an object knowledge if the underlying object has been removed, even if the knowledge hasn't yet been updated, however that's how it was done before the refactoring.