[simtests] fix TestDecCreateBreakdown, add orbat creation helpers
The motivation for FindOrCreateEntities was to have a simpler and
clearer way to create pieces of orbat. Create a tree made of
Party/Formation/Automat/Unit and call FindOrCreateEntities with it.
Requested entities will be either matched by name or created, and the
instances bound to the input data "Entity" field.

While the syntax is still a bit verbose, I see the following advantages:
- The desired structure is obvious in the code
- All the error handling is done by FindOrCreateEntities
- By defining types from the name instead of identifiers, it is much
  more legible
- We can extend it by adding attributes in user supplied structures.