[simtests] introduce $MASA_TIMEOUT_MULT to control timeouts
By default MASA_TIMEOUT_MULT=1, which defaults tests timeout to short
values usable on command line. When running gotester either locally with
multiple process or on Jenkins, it is forced to 10 to tolerate more load
on the host.

I used environment variables instead of command line options because I
found it easier to handle without having to share it in all tests
packages.