[simulation_kernel] make save and load symmetrical to reduce duplication
I strongly believe this 'pattern' should be applied everywhere.
Note the use of BOOST_CLASS_EXPORT_IMPLEMENT which forces template specialization thus making it possible to move the implementation of the serialize member function to the cpp.
BOOST_CLASS_EXPORT_KEY is not needed here because we don't have a derived class serialized through a pointer to a base class, we could have kept it though as it's harmless.