[svgl,clients_gui] fix SVG alpha
It's not possible to use SVG fillOpacity and strokeOpacity to handle the global layer-based alpha because nodes are allowed to override those values locally.
That was happening for several existing SVGs, for instance when drawing the object position for a 'Checkpoint' or a 'Rearm, Refuel and Resupply Point (R3P)'.
So instead we add a global opacity to multiply the local opacity with.