[simbench] add benchmarking tool
Run it like:

  $ go install masa/... && bin/simbench.exe -output results.xml
  OK: sw_pathfind_impossible: 2126.000000
  OK: sw_pathfind_long: 3116.000000
  OK: sw_pathfind_short: 101.000000
  writing results.xml

Tests can be filtered using -f, like -test.run.

The executable is configured to be run in sword project directory by
default, but the usual -application & co can be used to tell it where to
find the executable and data. The output file is meant be read by QA
benchmark runner and pushed in benchmark database.