[clients_kernel] tolerate iterating out of detection map (SWBUG-13322)
DetectionMapIterator is broken by construction and may end up generating
negative (or probably out of bound) coordinates, silently converted to
unsigned int causing overflows and triggering out of bound indexing. It
is a lot of work to rewrite this thing, in the meantime hide the problem
by detecting invalid lookups and returning default environment data to the
caller.