[clients_gui] simplify stencil buffer usage
Same idea as before: write in both the stencil and color buffers in one pass.
Also remove the contour drawn after the polygon filling, I suspect the idea was to try and make a smooth edge but this is unreliable and brittle at best (also it was probably no-op because the stencil buffer was still tested)
This is used for instance when drawing the convex hulls.