[gaming_app,preparation_app] remove symbol forced generation
No clue why it was there but it doesn't seem to be needed anymore as symbols are now generated synchronously on demand and take a negligible amount of time (tested on a whole army on Bertrand).
There is also no visible difference when starting the applications with and without this code.
Note that the symbols are still being generated after loading the exercise the first time they are needed, therefore the progress bar message as been left in place.