[simter] introduce future like mechanism to return pathfind results
The interface between TER_Pathfinder and its caller is brittle: they
create a DEC_PathComputer and poll it via GetState() before accessing
the computed path. There is no synchronization, you better be lucky to
read the correct path after watching GetState() change. While I am not
fond of "future" pattern, it perfectly fits the current architecture and
fixes the previous problem.

This commit introduce the API and use it in PathfindComputer. The
followup will address the other callers.