[frontend] create the autokill job on current process (SW-81575270)
The previous version was racy:
1- Spawn child process
2- Create an autokill job and assign it to it
But the process may stop between [1] and [2] causing the job creation
code to throw an error.

The new version creates an autokill job at selftraining startup and
assign selftraining job to it. By default, spawned process are bound to
their parent job, so any spawn process plus selftraining will belong the
job and will be terminated when the job is destroyed, that is when
selftraining terminates.

Also, failing to create the job no longer throws but log an error
instead.