[gaming][clients_gui] Removed T_CanBeRenamedFunctor from EntityImplementation
Rationale:
In many cases, the functor was not defined, thus wasting space. In most other cases, it was a std::function containing a closure object containing reference to profile_, which is less efficient than just storing a pointer to profile_.
A side effect is a cleanup of Team construction for the anonymous team.
TacticalLine_ABC destructor also fixed to avoid crash on disconnect.