[simter] rewrite "follow this path when possible" using A* heuristic
The previous version was editing waypoints to join a known itinerary as
soon as possible and returning precomputed path when moving between
known waypoints. The major problem with it were:
- It could backtrack on the itinerary because it was trying to join it
  at a point which was not directly accessible and had to move around
  some obstacle.
- Handling more than one itinerary was not obvious

The new version modifies the A* evaluation function to return the
minimum of the origin cost and another distance taking paths in account.
The latter is the sum of shortest distance of the node to the path, plus
the distance of the path to the destination, plus the distance to move
along the path with a modifier applied. It means we propagate at the
same time toward the destination or to a path nearby which would bring
us not too far from the destination. At the same time, the cost
function is minimized when moving on path edges.

Note the current implementation is terrible performance wise. The
point/edge distances computation are linear on path size and performed
for every evaluated node. Still it lets us prototype the feature, and
currently seems fast enough for a reduced usage like logistic or
engineer paths.