[gaming] fix objects destruction
Objects are registered in two different resolvers which both were more or less taking care of their life time in an ad hoc manner.
The fix introduces a TrackingResolver which combines a Resolver with an ElementObserver to automatically remove destroyed elements in order to factorize and properly implement this feature.