[addreports] add a -block argument, change the meaning of -count
-count is now the total number of reports to emit, while -block is the
number of reports to emit per client call. The idea is we want to
measure the amount of reports which can be created by tick. But the
reports creation call cannot span multiple tick, they are all emitted at
once. We have to find a parameter where we send enough calls to saturate
the system while not making each call long enough its heavily influences
the duration of the tick.