[simter] add prefered itineraries in dynamic graph (SWBUG-13293)
Itineraries can be created with arbitrary waypoints, and be made of
edges which do not exist in the graph. Weighting such edges is therefore
impossible.

Before computing the path, prefered itineraries are inserted in the
local dynamic graph, the path is computed and the itineraries are
removed. Note it also fixes unexplained unmatched segments in tests.