[simter] match itinerary edges by tagging dynamic data
Inserting dynamic data to ensure prefered itinerary edges belong to the
graph and can be weighted is a nice idea with one flaw: inserted
segments may be split in the dynamic graph and will not match against
the original one.

Instead of comparing with the original segments, we tag the inserted
edges with a new special linear type. When computing edge costs, we
check if the edge has this type. If so we consider it part of the
itinerary and forward the terrain type with the special bit remove to
the actual cost computer.

Unfortunately we cannot remove the segment matching part. Dynamic graph
terrain type merge rules are such than linear types are not merged but
overwritten, and only if it is not defined already. It means dynamic
edges matching existing roads never set the itinerary linear bit. In
this case fall back to geometry matching.

To summarize:
- The edge does not exist: create one with special type
- The edge exists: match segments

I still expect issues with existing edges which get split in the dynamic
graph and fail to match original segments.