[simter] turn dynamic data addition/removal into a single queue
Before this, ProcessDynamicData() was registering new dynamic data first
then applying the removal. Unfortunately, it means that, given a dynamic
data A:

  Add(A) + Add(A) + Remove(A) => nothing

because dynamic links are not referenced counted and the second
registration generated an empty set of "new" links. We had this pattern
when computing two paths one after the other in the same thread, using
the same itinerary. The sequence was like:

  Add(A1)
  ProcessDynamicData()
  ComputePath()
  Remove(A1)
  ...
  Add(A2)
  ProcessDynamicData() => applies Remove(A1), remove A1 and A2

We now process dynamic data operations in order.